fix bug when multiple shell commands are specified in RUN
If RUN is followed by two commands only the return value of second
determines whether RUN was successful or not. This may lead to cases
where docker build succeeds even when it should not.

For example
RUN /bin/false; /bin/true

would be considered a success.

Important thing to consider is that all yum/dnf commands should perform
clean in the same RUN operation if we wish to keep the container image
size small. Running clean in a second RUN operation leads to bigger
image size contrary to expectation because each operation leads to a
layer in the image.

With the above two points considered, I have replaced ";" with "&&"
where necessary and split a single RUN operation to two or many
operations in other places.

Signed-off-by: Raghavendra Talur <rtalur@redhat.com>